requirements.txt
amqp==2.2.2
astroid==2.0.4
billiard==3.5.0.3
cachetools==2.1.0
celery==4.1.0
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.1
django-allauth==0.37.1
django-crispy-forms==1.7.0
django-mathjax==0.0.8
django-rest-auth==0.9.3
djangorestframework==3.8.2
google-api-python-client==1.7.4
google-auth==1.5.1
google-auth-httplib2==0.0.3
gunicorn==19.7.1
html5lib==1.0b10
httplib2==0.10.3
idna==2.6
isort==4.3.4
kombu==4.1.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==3.1.0
numpy==1.12.1
oauth2client==4.1.3
oauthlib==2.1.0
olefile==0.44
pandas==0.20.3
Pillow==4.2.1
psycopg2==2.7.3.2
pyasn1==0.4.4
pyasn1-modules==0.2.2
pylint==2.1.1
PyPDF2==1.26.0
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2017.2
redis==2.10.6
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==1.0.0
rsa==4.0
scikit-learn==0.18.1
scipy==0.19.0
six==1.10.0
typed-ast==1.1.0
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
webencodings==0.5.1
whitenoise==3.3.0
wrapt==1.10.11
xhtml2pdf==0.2b1
XlsxWriter==1.0.2
